+++ products/bluebell/playground/src/main.rs
@@ -1,5 +1,8 @@
 mod app;
 mod state;
+mod vm_remote;
+mod bytecode_view;
+mod machine_view;
 
 use app::App;
